+++ dacite/src/lib.rs
@@ -58,4 +58,4 @@ pub const DACITE_API_VERSION: core::Version = core::Version {
 
 pub const DACITE_API_VERSION_MAJOR: u32 = 1;
 pub const DACITE_API_VERSION_MINOR: u32 = 0;
-pub const DACITE_API_VERSION_PATCH: u32 = 12;
+pub const DACITE_API_VERSION_PATCH: u32 = 13;